Move formatBytes to libHelpers
